Move PropertyAccessInfo classes to another file
